requirements.txt
python-binance==1.0.28
streamlit==1.28.0
pandas==2.1.0
python-dotenv==1.0.0
requests==2.31.0
loguru==0.7.2
plotly==5.15.0
